requirements.txt
pandas
requests
sqlalchemy
pymysql
scikit-learn
